requirements.txt
heatshrink2==0.13.0
Pillow==10.3.0
